+++ src/main/java/com/vendingMachine/model/dto/PurchaseResponse.java
@@ -4,11 +4,13 @@ import com.vendingMachine.model.entity.Product;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class PurchaseResponse {
 
     private Product product;
 
-    private Double userAmountLeft;
+    private BigDecimal userAmountLeft;
 }
